+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 ## [Unreleased]
 ### Fixed
 - Fixes a crash on track change when the plugin omits the `year` field for tracks without a year tag.
+- Fixes a home screen widget crash when tapping the widget to open the app, by launching the activity directly instead of routing through a Glance action callback.
 
 ## [1.6.0-rc.3] - 2026-04-19
 ### Fixed